requirements.txt
Django==3.1.7
djangorestframework==3.12.2
social-auth-app-django==4.0.0
cryptography==2.8
psycopg2==2.8.6
